requirements.txt
asana>=5.0.3
pandas>=2.2.0
pyarrow>=15.0.0
pyodbc>=5.0.1
python-dotenv>=1.0.1
setuptools
sqlalchemy>=2.0.25